requirements.txt
aioredis==1.2.0
amqp==2.3.2
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.11.1
Automat==0.7.0
billiard==3.5.0.4
celery==4.2.1
certifi==2018.8.24
cffi==1.11.5
channels==2.1.5
channels-redis==2.3.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.3.1
daphne==2.2.3
Django==2.1.2
django-redis==4.9.0
djangorestframework==3.9.0
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
kombu==4.2.1
lxml==4.2.5
msgpack==0.5.6
pycparser==2.19
PyHamcrest==1.9.0
PyMySQL==0.9.2
pytz==2018.5
redis==2.10.6
requests==2.19.1
six==1.11.0
Twisted==18.9.0
txaio==18.8.1
urllib3==1.23
vine==1.1.4
zope.interface==4.5.0
